pyproject.toml
[project]
name = "resume-shortlister"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "imbalanced-learn>=0.14.1",
    "joblib>=1.5.3",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "optuna>=4.7.0",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "selenium>=4.41.0",
    "sentence-transformers>=5.2.3",
    "shap>=0.51.0",
    "spacy>=3.8.11",
    "streamlit>=1.55.0",
    "tqdm>=4.67.3",
    "wordcloud>=1.9.6",
    "xgboost>=3.2.0",
]
